refactor: criação de pasta aluno para separação das models
