+++ src/main/java/br/org/institutobushido/controllers/AlunoControllers.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import br.org.institutobushido.model.Aluno;
+import br.org.institutobushido.model.aluno.Aluno;
 import br.org.institutobushido.services.AlunoServices;
 
 @RestController
