+++ src/main/java/br/org/institutobushido/services/AlunoServices.java
@@ -3,7 +3,7 @@ package br.org.institutobushido.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import br.org.institutobushido.model.Aluno;
+import br.org.institutobushido.model.aluno.Aluno;
 import br.org.institutobushido.repositories.AlunoRepositorio;
 import br.org.institutobushido.services.interfaces.AlunoServicesInterface;
