+++ src/main/java/br/org/institutobushido/services/interfaces/AlunoServicesInterface.java
@@ -1,6 +1,6 @@
 package br.org.institutobushido.services.interfaces;
 
-import br.org.institutobushido.model.Aluno;
+import br.org.institutobushido.model.aluno.Aluno;
 
 public interface AlunoServicesInterface {
     Aluno adicionarAluno(Aluno aluno);
